feat: Change client to stargate version and test code segmentation
