+++ build.gradle
@@ -51,7 +51,7 @@ idea {
 }
 
 group 'org.medibloc.panacea'
-version '0.1.5'
+version '2.0.0'
 
 sourceCompatibility = JavaVersion.VERSION_1_6
 targetCompatibility = JavaVersion.VERSION_1_6
