+++ src/test/java/org/medibloc/panacea/TestConst.java
@@ -1,0 +1,8 @@
+package org.medibloc.panacea;
+
+public class TestConst {
+
+    public static final String ownerMnemonic = "genuine key outside escape oval unhappy mansion cricket practice quarter purchase picnic layer bicycle stem soup column creek convince obey rather rice there alcohol";
+    public static final String toMnemonic = "margin river antenna nest drama combine crystal clarify route country genuine acoustic hold ocean image fox flame invite scrub chalk destroy resource item final";
+    public static final String denom = "umed";
+}
